+++ mobile/android/android-components/components/lib/crash/src/main/java/mozilla/components/lib/crash/Crash.kt
@@ -14,9 +14,12 @@ private const val INTENT_CRASH = "mozilla.components.lib.crash.CRASH"
 // Uncaught exception crash intent extras
 private const val INTENT_EXCEPTION = "exception"
 
+// Breadcrumbs intent extras
+private const val INTENT_BREADCRUMBS = "breadcrumbs"
+
 // Native code crash intent extras (Mirroring GeckoView values)
 private const val INTENT_MINIDUMP_PATH = "minidumpPath"
-private const val INTENT_EXTEAS_PATH = "extrasPath"
+private const val INTENT_EXTRAS_PATH = "extrasPath"
 private const val INTENT_MINIDUMP_SUCCESS = "minidumpSuccess"
 private const val INTENT_FATAL = "fatal"
 
@@ -30,15 +33,18 @@ sealed class Crash {
      * @property throwable The [Throwable] that caused the crash.
      */
     data class UncaughtExceptionCrash(
-        val throwable: Throwable
+        val throwable: Throwable,
+        val breadcrumbs: ArrayList<Breadcrumb>
     ) : Crash() {
         override fun toBundle() = Bundle().apply {
             putSerializable(INTENT_EXCEPTION, throwable as Serializable)
+            putParcelableArrayList(INTENT_BREADCRUMBS, breadcrumbs)
         }
 
         companion object {
             internal fun fromBundle(bundle: Bundle) = UncaughtExceptionCrash(
-                bundle.getSerializable(INTENT_EXCEPTION) as Throwable
+                bundle.getSerializable(INTENT_EXCEPTION) as Throwable,
+                bundle.getParcelableArrayList(INTENT_BREADCRUMBS) ?: arrayListOf()
             )
         }
     }
@@ -60,21 +66,24 @@ sealed class Crash {
         val minidumpPath: String,
         val minidumpSuccess: Boolean,
         val extrasPath: String,
-        val isFatal: Boolean
+        val isFatal: Boolean,
+        val breadcrumbs: ArrayList<Breadcrumb>
     ) : Crash() {
         override fun toBundle() = Bundle().apply {
             putString(INTENT_MINIDUMP_PATH, minidumpPath)
             putBoolean(INTENT_MINIDUMP_SUCCESS, minidumpSuccess)
-            putString(INTENT_EXTEAS_PATH, extrasPath)
+            putString(INTENT_EXTRAS_PATH, extrasPath)
             putBoolean(INTENT_FATAL, isFatal)
+            putParcelableArrayList(INTENT_BREADCRUMBS, breadcrumbs)
         }
 
         companion object {
             internal fun fromBundle(bundle: Bundle) = NativeCodeCrash(
                 bundle.getString(INTENT_MINIDUMP_PATH, ""),
                 bundle.getBoolean(INTENT_MINIDUMP_SUCCESS, false),
-                bundle.getString(INTENT_EXTEAS_PATH, ""),
-                bundle.getBoolean(INTENT_FATAL, false)
+                bundle.getString(INTENT_EXTRAS_PATH, ""),
+                bundle.getBoolean(INTENT_FATAL, false),
+                bundle.getParcelableArrayList(INTENT_BREADCRUMBS) ?: arrayListOf()
             )
         }
     }