+++ mobile/android/android-components/components/lib/crash/src/main/java/mozilla/components/lib/crash/handler/ExceptionHandler.kt
@@ -27,7 +27,8 @@ class ExceptionHandler(
         try {
             crashing = true
 
-            crashReporter.onCrash(context, Crash.UncaughtExceptionCrash(throwable))
+            crashReporter.onCrash(context, Crash.UncaughtExceptionCrash(throwable,
+                    crashReporter.crashBreadcrumbs))
 
             defaultExceptionHandler?.uncaughtException(thread, throwable)
         } finally {